Reworked PacManComponent
Moved all the movement logic inside the PacManComponent into a new CharacterComponent. This is as anticipation of the implementation of the ghosts. The ghost GameObjects will then also be composed of this CharacterComponent to handle their movement and a GhostComponent to decide in which direction to set their CharacterComponent.

Added a method to the SceneManager class to get the active scene.

Added a method to the Scene class to find and return an object by name.