Reworked MapComponent & PacManComponent
Reworked MapComponent. Instead of having a separate method to check the type of a cell at a certain position for each type of cell, there is now one method that just returns the type of the cell at a position.

The player will now not be able to move anymore once they loses all of their lives