Minor changes
Added IsActive tag to GameObjects. Only active objects get updated and rendered.

Added UI for entering a name and saving score.

Added TextComponent showing the appropriate text, to the GameMode GameObject that gets displayed when the game is over.

Reworked gameplay so that the pacman and the ghosts no longer interact when the game is over.